+++ gr00t/model/policy.py
@@ -28,7 +28,7 @@ from gr00t.data.embodiment_tags import EmbodimentTag
 from gr00t.data.schema import DatasetMetadata
 from gr00t.data.transform.base import ComposedModalityTransform
 from gr00t.model.gr00t_n1 import GR00T_N1
-from gr00t.model.pretrained import Gr00tMixin, ModalityConfigDict
+from gr00t.utils.pretrained import Gr00tMixin, ModalityConfigDict
 
 COMPUTE_DTYPE = torch.bfloat16
